structural changes to address tracebacks
