+++ gluster-colonizer.py
@@ -179,7 +179,8 @@ if nm_storage_interface.startswith("bond-") or nm_storage_interface.startswith("
 else:
     storage_interface = nm_storage_interface
 
-setupLogging(args)
+logfile = 'gluster-colonizer.log'
+logger = setupLogging(args, logfile)
 
 # NOTE: Moved below to g1modules.py
 ## Init logging to log to console screen and file
@@ -244,6 +245,60 @@ setupLogging(args)
 #        else:
 #            print "Please enter either 'yes' or 'no'\r\n"
 
+#def stopDhcpService():
+#    # Function to stop specialized DHCP server
+#    killDnsmasq()
+#    host_command('/bin/firewall-cmd --remove-service=dhcp')
+#    host_command('/bin/nmcli con reload %s' % nm_mgmt_interface)
+#    host_command('/bin/nmcli con up %s' % nm_mgmt_interface)
+
+#def killDnsmasq():
+#    # Function to stop any existing dnsmasq processes
+#    logger.debug("Killing any existing dnsmasq processes")
+#    p1 = Popen(shlex.split('ps -e'), stdout=PIPE)
+#    p2 = Popen(
+#        shlex.split('grep dnsmasq'),
+#        stdin=p1.stdout,
+#        stdout=PIPE,
+#        stderr=STDOUT)
+#    pOut, _ = p2.communicate()
+#    for line in pOut.splitlines():
+#        if 'dnsmasq' in line:
+#            pid = int(line.split(None, 1)[0])
+#            os.kill(pid, signal.SIGKILL)
+#    logger.debug("Wiping the dnsmasq.leases file")
+#    host_command("echo '' > /var/lib/dnsmasq/dnsmasq.leases", shell=True)
+
+#def host_command(command, shell=False):
+#    # Function to execute system commands
+#    if shell == True:
+#        cmd_args = command
+#    else:
+#        cmd_args = shlex.split(command)
+#
+#    logger.debug("Initiating Subprocess: " + command)
+#
+#    try:
+#        cmd_proc = Popen(
+#            cmd_args,
+#            stdout=PIPE,
+#            stderr=STDOUT,
+#            universal_newlines=True,
+#            shell=shell)
+#
+#        proc_output, _ = cmd_proc.communicate()
+#
+#        if proc_output.strip() != "":
+#            logger.debug("Subprocess output: " + proc_output)
+#    except (OSError, CalledProcessError) as exception:
+#        logger.error("Subprocess exception occured: " + str(exception))
+#        abortSetup("Subprocess failed")
+#
+#    return proc_output
+
+
+
+
 def set_ha_node_count():
     ha_node_count = int(
         math.ceil(int(len(g1Hosts)) / float(ha_node_factor)))
@@ -285,34 +340,6 @@ def ipValidator(user_message,
         consumed_ips.append(str(ip))
         break
     return str(ip)
-
-
-def host_command(command, shell=False):
-    # Function to execute system commands
-    if shell == True:
-        cmd_args = command
-    else:
-        cmd_args = shlex.split(command)
-
-    logger.debug("Initiating Subprocess: " + command)
-
-    try:
-        cmd_proc = Popen(
-            cmd_args,
-            stdout=PIPE,
-            stderr=STDOUT,
-            universal_newlines=True,
-            shell=shell)
-
-        proc_output, _ = cmd_proc.communicate()
-
-        if proc_output.strip() != "":
-            logger.debug("Subprocess output: " + proc_output)
-    except (OSError, CalledProcessError) as exception:
-        logger.error("Subprocess exception occured: " + str(exception))
-        abortSetup("Subprocess failed")
-
-    return proc_output
 
 
 def run_ansible_playbook(playbook, continue_on_fail=False, become=False, askConnPass=False, askSudoPass=False):
@@ -370,24 +397,6 @@ def run_ansible_playbook(playbook, continue_on_fail=False, become=False, askConn
     return True
 
 
-def killDnsmasq():
-    # Function to stop any existing dnsmasq processes
-    logger.debug("Killing any existing dnsmasq processes")
-    p1 = Popen(shlex.split('ps -e'), stdout=PIPE)
-    p2 = Popen(
-        shlex.split('grep dnsmasq'),
-        stdin=p1.stdout,
-        stdout=PIPE,
-        stderr=STDOUT)
-    pOut, _ = p2.communicate()
-    for line in pOut.splitlines():
-        if 'dnsmasq' in line:
-            pid = int(line.split(None, 1)[0])
-            os.kill(pid, signal.SIGKILL)
-    logger.debug("Wiping the dnsmasq.leases file")
-    host_command("echo '' > /var/lib/dnsmasq/dnsmasq.leases", shell=True)
-
-
 def startDhcpService():
     # Function to set and initiate services for this node as the deployment master
     global mgmt_subnet
@@ -430,14 +439,6 @@ def startDhcpService():
     # TODO: Discuss subnet for initial DHCP config
     host_command('/sbin/dnsmasq --interface=%s --dhcp-range=%s,%s,12h' %
                  (mgmt_interface, mgmt_subnet[2], mgmt_subnet[-2]))
-
-
-def stopDhcpService():
-    # Function to stop specialized DHCP server
-    killDnsmasq()
-    host_command('/bin/firewall-cmd --remove-service=dhcp')
-    host_command('/bin/nmcli con reload %s' % nm_mgmt_interface)
-    host_command('/bin/nmcli con up %s' % nm_mgmt_interface)
 
 
 def collectDeploymentInformation():
